# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: COIMAGEN OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /costs:
    get:
      operationId: listCosts
      tags: [costs]
      parameters:
        - { name: month, in: query, required: false, schema: { type: string } }
        - { name: category, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cost"
    post:
      operationId: createCost
      tags: [costs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cost"
  /costs/{id}:
    get:
      operationId: getCost
      tags: [costs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cost"
        "404":
          description: Not found
    patch:
      operationId: updateCost
      tags: [costs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CostUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cost"
        "404":
          description: Not found
    delete:
      operationId: deleteCost
      tags: [costs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /costs/summary/{month}:
    get:
      operationId: getCostSummary
      tags: [costs]
      parameters:
        - { name: month, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostSummary"

  # ─── AUDIT LOG ─────────────────────────────────────────────────────────────
...
  /daily-sprints:
    get:
      operationId: listDailySprints
      tags: [daily-sprints]
      summary: List daily sprints
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailySprint"
    post:
      operationId: createDailySprint
      tags: [daily-sprints]
      summary: Create daily sprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailySprintCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySprint"
  /daily-sprints/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      operationId: getDailySprint
      tags: [daily-sprints]
      summary: Get daily sprint
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySprint"
    patch:
      operationId: updateDailySprint
      tags: [daily-sprints]
      summary: Update daily sprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailySprintUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySprint"
    delete:
      operationId: deleteDailySprint
      tags: [daily-sprints]
      summary: Delete daily sprint
      responses:
        "204":
          description: Deleted

  # ─── ROADMAP ─────────────────────────────────────────────────────────────────
...
  /contracts:
    get:
      operationId: listContracts
      summary: List contracts
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: clientId
          schema: { type: integer }
        - in: query
          name: projectId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Contract" }
    post:
      operationId: createContract
      summary: Create contract
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContractCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
  /contracts/{id}:
    get:
      operationId: getContract
      summary: Get contract by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
    patch:
      operationId: updateContract
      summary: Update contract
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContractUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contract" }
    delete:
      operationId: deleteContract
      summary: Delete contract
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No Content
...
components:
...
  schemas:
...
    Cost:
      type: object
      required: [id, category, month, amount, createdAt]
      properties:
        id: { type: integer }
        category: { type: string, enum: [openai, claude, gemini, whatsapp, n8n, hosting, replit, other] }
        month: { type: string }
        amount: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    CostInput:
      type: object
      required: [category, month, amount]
      properties:
        category: { type: string, enum: [openai, claude, gemini, whatsapp, n8n, hosting, replit, other] }
        month: { type: string }
        amount: { type: number }
        notes: { type: string }
    CostUpdate:
      type: object
      properties:
        category: { type: string, enum: [openai, claude, gemini, whatsapp, n8n, hosting, replit, other] }
        month: { type: string }
        amount: { type: number }
        notes: { type: string }
    CostSummary:
      type: object
      required: [month, totalCosts, totalRevenue, estimatedMargin, breakdown]
      properties:
        month: { type: string }
        totalCosts: { type: number }
        totalRevenue: { type: number }
        estimatedMargin: { type: number }
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/CostBreakdownItem"
    CostBreakdownItem:
      type: object
      required: [category, amount, percentage]
      properties:
        category: { type: string }
        amount: { type: number }
        percentage: { type: number }

    # ── AUDIT LOG ─────────────────────────────────────────────────────────────
...
    DailySprint:
      type: object
      required: [id, date, createdAt]
      properties:
        id: { type: integer }
        date: { type: string }
        objective: { type: ["string", "null"] }
        doneYesterday: { type: ["string", "null"] }
        replitWorking: { type: ["string", "null"] }
        todayPlan: { type: ["string", "null"] }
        blockers: { type: ["string", "null"] }
        deliverables: { type: ["string", "null"] }
        result: { type: ["string", "null"] }
        tomorrowPending: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    DailySprintCreate:
      type: object
      required: [date]
      properties:
        date: { type: string }
        objective: { type: string }
        doneYesterday: { type: string }
        replitWorking: { type: string }
        todayPlan: { type: string }
        blockers: { type: string }
        deliverables: { type: string }
        result: { type: string }
        tomorrowPending: { type: string }
        status: { type: string }
    DailySprintUpdate:
      type: object
      properties:
        date: { type: string }
        objective: { type: string }
        doneYesterday: { type: string }
        replitWorking: { type: string }
        todayPlan: { type: string }
        blockers: { type: string }
        deliverables: { type: string }
        result: { type: string }
        tomorrowPending: { type: string }
        status: { type: string }

    # ─── ROADMAP ──────────────────────────────────────────────────────────────
...
    Contract:
      type: object
      required: [id, type, status, title, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        status: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        service: { type: ["string", "null"] }
        clientId: { type: ["integer", "null"] }
        projectId: { type: ["integer", "null"] }
        workflowId: { type: ["integer", "null"] }
        invoiceId: { type: ["integer", "null"] }
        approvalId: { type: ["integer", "null"] }
        content: { type: ["string", "null"] }
        amount: { type: ["integer", "null"] }
        currency: { type: ["string", "null"] }
        terms: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"] }
        signedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
        createdBy: { type: ["string", "null"] }
        signedBy: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    ContractCreate:
      type: object
      required: [type, title]
      properties:
        type: { type: string }
        title: { type: string }
        description: { type: string }
        service: { type: string }
        clientId: { type: integer }
        projectId: { type: integer }
        workflowId: { type: integer }
        invoiceId: { type: integer }
        approvalId: { type: integer }
        content: { type: string }
        amount: { type: integer }
        currency: { type: string }
        terms: { type: string }
        notes: { type: string }
        sentAt: { type: string }
        signedAt: { type: string }
        expiresAt: { type: string }
        createdBy: { type: string }
        signedBy: { type: string }
    ContractUpdate:
      type: object
      properties:
        type: { type: string }
        status: { type: string }
        title: { type: string }
        description: { type: string }
        service: { type: string }
        clientId: { type: integer }
        projectId: { type: integer }
        workflowId: { type: integer }
        invoiceId: { type: integer }
        approvalId: { type: integer }
        content: { type: string }
        amount: { type: integer }
        currency: { type: string }
        terms: { type: string }
        notes: { type: string }
        sentAt: { type: string }
        signedAt: { type: string }
        expiresAt: { type: string }
        createdBy: { type: string }
        signedBy: { type: string }
...